Create a @Component annotated interface to enable injection to our activity
It assigns references in our activities, services, or fragments
to have access to singletons we earlier defined.

*Make sure to make 'Build' for the generation of the Dagger%yourcomponentsname%
class. That class will be responsible for instantiating an instance of
our dependency graph and using it to perform the injection work for fields
annotated with @Inject.

The instantiation of this class should be done in a custom Application class
since this will be used in the whole project.